Create a styles file for each individual screen and each component file
