+++ app/limitaciongastos.tsx
@@ -5,7 +5,7 @@ import { collection, getDocs, doc, updateDoc } from "firebase/firestore";
 import { database } from "../src/config/firebase";
 import { Ionicons } from "@expo/vector-icons";
 import { router } from "expo-router";
-import { style } from "@/styles/style";
+import { limitacionGastos } from "@/styles/limitacion-gastos";
 
 type Categoria = {
     label: string;
@@ -65,28 +65,28 @@ export default function LimitacionGastos() {
     };
 
     return (
-        <SafeAreaView style={style.limitContainer}>
+        <SafeAreaView style={limitacionGastos.limitContainer}>
 
             {/* Back */}
-            <TouchableOpacity style={style.backButton} onPress={() => router.push("/configuracion")}>
+            <TouchableOpacity style={limitacionGastos.backButton} onPress={() => router.push("/configuracion")}>
                 <Ionicons name="arrow-back" size={26} color="black" />
             </TouchableOpacity>
 
             {/* Título */}
-            <Text style={style.limitTitle}>Limitación de Gastos</Text>
+            <Text style={limitacionGastos.limitTitle}>Limitación de Gastos</Text>
 
             {/* Tarjeta */}
-            <View style={style.limitCard}>
+            <View style={limitacionGastos.limitCard}>
 
                 <View style={{ marginBottom: 20 }}>
-                    <Text style={style.limitLabel}>Selecciona una categoría</Text>
+                    <Text style={limitacionGastos.limitLabel}>Selecciona una categoría</Text>
 
                     {/* Botón que abre/cierra el dropdown */}
                     <TouchableOpacity
-                        style={style.dropdownButton}
+                        style={limitacionGastos.dropdownButton}
                         onPress={() => setShowDropdown(!showDropdown)}
                     >
-                        <Text style={style.dropdownButtonText}>
+                        <Text style={limitacionGastos.dropdownButtonText}>
                             {categorias.find((c) => c.value === categoriaSeleccionada)?.label || "Seleccionar categoría"}
                         </Text>
                         <Ionicons name={showDropdown ? "chevron-up" : "chevron-down"} size={22} color="#555" />
@@ -94,18 +94,18 @@ export default function LimitacionGastos() {
 
                     {/* Lista desplegable */}
                     {showDropdown && (
-                        <View style={style.dropdownContainer}>
+                        <View style={limitacionGastos.dropdownContainer}>
                             <ScrollView style={{ maxHeight: 200 }}>
                                 {categorias.map((cat) => (
                                     <TouchableOpacity
                                         key={cat.value}
-                                        style={style.dropdownItem}
+                                        style={limitacionGastos.dropdownItem}
                                         onPress={() => {
                                             setCategoriaSeleccionada(cat.value);
                                             setShowDropdown(false);
                                         }}
                                     >
-                                        <Text style={style.dropdownItemText}>{cat.label}</Text>
+                                        <Text style={limitacionGastos.dropdownItemText}>{cat.label}</Text>
                                     </TouchableOpacity>
                                 ))}
                             </ScrollView>
@@ -115,18 +115,18 @@ export default function LimitacionGastos() {
                 </View>
 
 
-                <Text style={style.limitLabel}>Límite de gasto (MXN)</Text>
+                <Text style={limitacionGastos.limitLabel}>Límite de gasto (MXN)</Text>
 
                 <TextInput
                     value={limite}
                     onChangeText={setLimite}
                     keyboardType="numeric"
                     placeholder="Ejemplo: 1000"
-                    style={style.limitInput}
+                    style={limitacionGastos.limitInput}
                 />
 
-                <TouchableOpacity style={style.limitBtnGuardar} onPress={guardarLimite}>
-                    <Text style={style.limitTxtGuardar}>Guardar límite</Text>
+                <TouchableOpacity style={limitacionGastos.limitBtnGuardar} onPress={guardarLimite}>
+                    <Text style={limitacionGastos.limitTxtGuardar}>Guardar límite</Text>
                 </TouchableOpacity>
             </View>
