+++ app/vertodo.tsx
@@ -1,5 +1,5 @@
-import TodosRegistros from "@/components/todos";
-import { style } from "@/styles/style";
+import TodosRegistros from "@/components/Todos";
+import { vertodo } from "@/styles/vertodo";
 import Ionicons from "@expo/vector-icons/Ionicons";
 import { router } from "expo-router";
 import { ScrollView, TouchableOpacity, View } from "react-native";
@@ -10,7 +10,7 @@ export default function(){
         <SafeAreaView>
             <View>
                 <View>
-                    <TouchableOpacity style={style.backButton} onPress={() => router.push("/inicio")}>
+                    <TouchableOpacity style={vertodo.backButton} onPress={() => router.push("/inicio")}>
                 <Ionicons name="arrow-back" size={26} color="black" />
             </TouchableOpacity>
                 </View>
