+++ components/recientes.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Text, View, FlatList, TouchableOpacity } from "react-native";
 import { collection, onSnapshot, query, where, orderBy, limit } from "firebase/firestore";
 import { auth, database } from "../src/config/firebase";
-import { style } from "@/styles/style";
+import { recientes } from "@/styles/recientes";
 import { router } from "expo-router";
 
 type Transaccion = {
@@ -49,14 +49,14 @@ export default function UltimosRegistros() {
 
     const renderItem = ({ item }: { item: Transaccion }) => (
         <TouchableOpacity
-            style={style.registroItem}
+            style={recientes.registroItem}
             onPress={() => router.push(`/detalles?id=${item.id}`)}
         >
-            <Text style={style.registroCategoria}>{item.categoria}</Text>
-            <Text style={style.registroDescripcion}>{item.descripcion}</Text>
-            <Text style={style.registroTipo}>{item.tipo}</Text>
-            <Text style={style.registroMonto}>${item.monto.toFixed(2)}</Text>
-            <Text style={style.registroFecha}>
+            <Text style={recientes.registroCategoria}>{item.categoria}</Text>
+            <Text style={recientes.registroDescripcion}>{item.descripcion}</Text>
+            <Text style={recientes.registroTipo}>{item.tipo}</Text>
+            <Text style={recientes.registroMonto}>${item.monto.toFixed(2)}</Text>
+            <Text style={recientes.registroFecha}>
                 {item.fecha?.toDate ? item.fecha.toDate().toLocaleDateString() : ""}
             </Text>
         </TouchableOpacity>
