--- styles/style.ts
@@ -1,804 +1,0 @@
-import { StyleSheet } from "react-native"
-
-export const style = StyleSheet.create({
-
-    container: {
-        flex: 1,
-        backgroundColor: "#F8F9FA",
-        padding: 25,
-    },
-
-    title: {
-        fontSize: 26,
-        color: "#212529",
-        fontWeight: "bold",
-        marginBottom: 20,
-    },
-
-    formTitle: {
-        fontSize: 22,
-        fontWeight: "600",
-        color: "#212529",
-        marginBottom: 20,
-        textAlign: "center",
-    },
-
-    label: {
-        fontSize: 16,
-        color: "#495057",
-        marginBottom: 6,
-        fontWeight: "600",
-    },
-
-    itemText: {
-        fontSize: 18,
-        color: "#495057",
-        marginBottom: 10,
-    },
-
-    separator: {
-        marginVertical: 10,
-        fontSize: 16,
-        color: "#495057",
-        textAlign: "center",
-    },
-
-    formContainer: {
-        marginBottom: 40,
-    },
-
-    inputGroup: {
-        marginBottom: 20,
-    },
-
-    input: {
-        backgroundColor: "#FFFFFF",
-        borderRadius: 10,
-        padding: 14,
-        fontSize: 16,
-        borderWidth: 1,
-        borderColor: "#DEE2E6",
-        color: "#212529",
-    },
-
-    eyeButton: {
-        position: "absolute",
-        right: 15,
-        top: 14,
-    },
-
-    card: {
-        backgroundColor: "white",
-        padding: 20,
-        borderRadius: 15,
-        elevation: 3,
-        shadowColor: "#000",
-        shadowOpacity: 0.1,
-        shadowRadius: 5,
-        shadowOffset: { width: 0, height: 3 },
-        marginTop: 10,
-    },
-
-    buttonsContainer: {
-        alignItems: "center",
-    },
-
-    btnPrimary: {
-        backgroundColor: "#4C6EF5",
-        paddingVertical: 14,
-        paddingHorizontal: 25,
-        borderRadius: 12,
-        width: "100%",
-        marginBottom: 10,
-        alignItems: "center",
-    },
-
-    btnSecondary: {
-        backgroundColor: "#339AF0",
-        paddingVertical: 14,
-        paddingHorizontal: 25,
-        borderRadius: 12,
-        width: "100%",
-        alignItems: "center",
-    },
-
-    btnText: {
-        fontSize: 17,
-        fontWeight: "600",
-        color: "white",
-    },
-
-    backButton: {
-        width: 40,
-        height: 40,
-        borderRadius: 10,
-        backgroundColor: "#E9ECEF",
-        justifyContent: "center",
-        alignItems: "center",
-        marginBottom: 15,
-    },
-   
-
-    cargandoContainer: {
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-    },
-
-    cargandoTexto: {
-        fontSize: 18,
-    },
-    registroContainer: {
-        flex: 1,
-        padding: 20,
-        backgroundColor: "#F5F5F5",
-    },
-
-    campoContainer: {
-        marginBottom: 15,
-    },
-
-    labelTexto: {
-        fontSize: 16,
-        fontWeight: "600",
-        marginBottom: 5,
-        color: "#333",
-    },
-
-    inputRegistro: {
-        backgroundColor: "#FFF",
-        padding: 12,
-        borderRadius: 10,
-        borderWidth: 1,
-        borderColor: "#DDD",
-        fontSize: 15,
-        elevation: 2,
-    },
-
-    eyeButtonRegistro: {
-        position: "absolute",
-        right: 15,
-        top: 38,
-    },
-
-    textoAdicional: {
-        fontSize: 12,
-        marginTop: 5,
-        color: "#666",
-    },
-
-    checkboxContainerRegistro: {
-        flexDirection: "row",
-        alignItems: "center",
-        marginTop: 10,
-        marginBottom: 20,
-    },
-
-    textcheckRegistro: {
-        fontSize: 14,
-        color: "#333",
-    },
-
-    btnRegistrar: {
-        backgroundColor: "#007AFF",
-        padding: 15,
-        borderRadius: 12,
-        alignItems: "center",
-    },
-
-    txtBtnRegistrar: {
-        fontSize: 18,
-        fontWeight: "bold",
-        color: "#FFF",
-    },
-    transContainer: {
-        flex: 1,
-        padding: 20,
-        backgroundColor: "#F5F5F5",
-    },
-
-    tituloTrans: {
-        fontSize: 22,
-        fontWeight: "bold",
-        textAlign: "center",
-        marginBottom: 15,
-        color: "#333",
-    },
-
-    campoTrans: {
-        marginBottom: 18,
-    },
-
-    labelTrans: {
-        fontSize: 16,
-        fontWeight: "600",
-        marginBottom: 6,
-        color: "#333",
-    },
-
-    selectInput: {
-        backgroundColor: "#FFF",
-        padding: 12,
-        borderRadius: 10,
-        borderWidth: 1,
-        borderColor: "#DDD",
-        elevation: 2,
-        flex: 1,
-        fontSize: 15,
-    },
-
-    selectRow: {
-        flexDirection: "row",
-        alignItems: "center",
-    },
-
-    listaOpciones: {
-        backgroundColor: "#FFF",
-        marginTop: 5,
-        paddingVertical: 10,
-        borderRadius: 10,
-        elevation: 2,
-    },
-
-    inputTrans: {
-        backgroundColor: "#FFF",
-        padding: 12,
-        borderRadius: 10,
-        borderWidth: 1,
-        borderColor: "#DDD",
-        fontSize: 15,
-        elevation: 2,
-    },
-
-    btnGuardar: {
-        backgroundColor: "#007AFF",
-        padding: 15,
-        borderRadius: 12,
-        alignItems: "center",
-        marginTop: 25,
-    },
-    btnCerrar: {
-        backgroundColor: "#ff0000ff",
-        padding: 15,
-        borderRadius: 12,
-        alignItems: "center",
-        marginTop: 25,
-    },
-
-    txtGuardar: {
-        color: "#FFF",
-        fontSize: 18,
-        fontWeight: "bold",
-    },
-    registroItem: {
-        padding: 16,
-        backgroundColor: "#ffffff",
-        borderRadius: 12,
-        marginBottom: 12,
-        borderWidth: 1,
-        borderColor: "#e6e6e6",
-        shadowColor: "#000",
-        shadowOpacity: 0.05,
-        shadowRadius: 4,
-        shadowOffset: { width: 0, height: 2 }
-    },
-
-    registroCategoria: {
-        fontSize: 16,
-        fontWeight: "600",
-        color: "#333"
-    },
-
-    registroDescripcion: {
-        fontSize: 14,
-        color: "#555",
-        marginBottom: 4
-    },
-
-    registroTipo: {
-        fontSize: 14,
-        fontWeight: "500",
-        color: "#000"
-    },
-
-    registroMonto: {
-        fontSize: 18,
-        fontWeight: "bold",
-        color: "#0A7F00",
-        marginTop: 4
-    },
-
-    registroFecha: {
-        fontSize: 12,
-        color: "#777",
-        marginTop: 4
-    },
-    btnEliminar: {
-        marginTop: 20,
-        alignItems: "center"
-    },
-    btnsettings: {
-
-    },
-    areausuario: {
-        padding: 20,
-        flex: 1,
-    },
-
-    loadingContainerusuario: {
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-    },
-
-    loadingTextusuario: {
-        fontSize: 18,
-    },
-
-    titleusuario: {
-        fontSize: 20,
-        fontWeight: "bold",
-        marginBottom: 15,
-    },
-
-    itemusuario: {
-        marginBottom: 8,
-        fontSize: 16,
-    },
-
-    labelusuario: {
-        fontWeight: "bold",
-    },
-    // --- ESTILOS DE LOGIN ---
-    loginHeader: {
-        marginBottom: 40,
-        alignItems: "center",
-    },
-
-    loginTitle: {
-        fontSize: 28,
-        fontWeight: "bold",
-        color: "#1B263B",
-        textAlign: "center",
-    },
-
-    loginSubtitle: {
-        fontSize: 15,
-        color: "#6C757D",
-        marginTop: 6,
-        textAlign: "center",
-    },
-
-    loginForm: {
-        marginBottom: 35,
-    },
-
-    loginInputGroup: {
-        marginBottom: 20,
-    },
-
-    loginLabel: {
-        fontSize: 15,
-        fontWeight: "600",
-        color: "#495057",
-        marginBottom: 6,
-    },
-
-    loginInput: {
-        backgroundColor: "#FFFFFF",
-        borderRadius: 12,
-        padding: 14,
-        fontSize: 16,
-        borderWidth: 1,
-        borderColor: "#D0D4DA",
-        color: "#212529",
-        elevation: 1,
-    },
-
-    loginEyeButton: {
-        position: "absolute",
-        right: 15,
-        top: 16,
-    },
-
-    loginButtons: {
-        alignItems: "center",
-        marginTop: 10,
-    },
-
-    loginBtnPrimary: {
-        backgroundColor: "#4C6EF5",
-        paddingVertical: 14,
-        borderRadius: 12,
-        width: "100%",
-        alignItems: "center",
-        marginBottom: 8,
-        elevation: 2,
-    },
-
-    loginBtnSecondary: {
-        backgroundColor: "#E9ECEF",
-        paddingVertical: 14,
-        borderRadius: 12,
-        width: "100%",
-        alignItems: "center",
-    },
-
-    loginBtnText: {
-        fontSize: 17,
-        fontWeight: "600",
-        color: "white",
-    },
-
-    loginBtnTextSecondary: {
-        fontSize: 17,
-        fontWeight: "600",
-        color: "#495057",
-    },
-
-    loginSeparator: {
-        marginVertical: 10,
-        fontSize: 15,
-        color: "#6C757D",
-    },
-
-    inicioContainer: {
-        flex: 1,
-        padding: 20,
-        backgroundColor: "#F0F2F5",
-    },
-
-    balanceContainer: {
-        marginBottom: 20,
-        padding: 20,
-        borderRadius: 16,
-        backgroundColor: "#0a7ea4",
-        shadowColor: "#000",
-        shadowOpacity: 0.15,
-        shadowRadius: 6,
-        shadowOffset: { width: 0, height: 4 },
-        elevation: 6,
-    },
-
-    balanceTitulo: {
-        fontSize: 26,
-        fontWeight: "900",
-        color: "#FFF",
-        textAlign: "center",
-    },
-
-    ingresosEgresosBox: {
-        backgroundColor: "#FFF",
-        padding: 20,
-        borderRadius: 16,
-        marginBottom: 20,
-        shadowColor: "#000",
-        shadowOpacity: 0.10,
-        shadowRadius: 5,
-        shadowOffset: { width: 0, height: 3 },
-        elevation: 4,
-    },
-
-    ingresoTexto: {
-        fontSize: 18,
-        fontWeight: "700",
-        color: "#2ECC71",
-        marginBottom: 6,
-    },
-
-    egresoTexto: {
-        fontSize: 18,
-        fontWeight: "700",
-        color: "#E74C3C",
-    },
-
-    chartContainer: {
-        backgroundColor: "#FFF",
-        padding: 20,
-        borderRadius: 16,
-        elevation: 4,
-        marginBottom: 25,
-        shadowColor: "#000",
-        shadowOpacity: 0.10,
-        shadowRadius: 6,
-        shadowOffset: { width: 0, height: 3 },
-    },
-
-    chartTitulo: {
-        textAlign: "center",
-        marginBottom: 12,
-        fontWeight: "800",
-        fontSize: 18,
-        color: "#2C3E50",
-    },
-
-    recientesScroll: {
-        marginTop: 10,
-        marginBottom: 20,
-    },
-
-    recientesTitulo: {
-        fontSize: 20,
-        fontWeight: "800",
-        marginBottom: 12,
-        color: "#2C3E50",
-    },
-
-    verMasContainer: {
-        marginTop: 5,
-        alignItems: "center",
-    },
-
-    verMasTexto: {
-        fontSize: 16,
-        fontWeight: "600",
-        color: "#4C6EF5",
-    },
-
-    botonFlotanteContainer: {
-        position: "absolute",
-        bottom: 25,
-        right: 25,
-        backgroundColor: "#0a7ea4",
-        width: 65,
-        height: 65,
-        borderRadius: 40,
-        justifyContent: "center",
-        alignItems: "center",
-        shadowColor: "#000",
-        shadowOpacity: 0.25,
-        shadowRadius: 6,
-        shadowOffset: { width: 0, height: 3 },
-        elevation: 6,
-    },
-    // -------- ESTILOS DE CONFIGURACIÓN -------- //
-configContainer: {
-    flex: 1,
-    backgroundColor: "#F5F5F5",
-    padding: 20,
-},
-
-configCard: {
-    backgroundColor: "#FFF",
-    padding: 20,
-    borderRadius: 15,
-    elevation: 3,
-    shadowColor: "#000",
-    shadowOpacity: 0.1,
-    shadowRadius: 5,
-    shadowOffset: { width: 0, height: 3 },
-},
-
-configTitle: {
-    fontSize: 24,
-    fontWeight: "800",
-    color: "#1B263B",
-    marginBottom: 25,
-    textAlign: "center",
-},
-
-configButton: {
-    backgroundColor: "#4C6EF5",
-    paddingVertical: 14,
-    borderRadius: 12,
-    marginBottom: 15,
-    alignItems: "center",
-},
-
-configButtonText: {
-    color: "white",
-    fontSize: 16,
-    fontWeight: "600",
-},
-// --- ESTILOS DE AGREGAR CATEGORÍA ---
-catContainer: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: "#F5F5F5",
-},
-
-catHeader: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginBottom: 25,
-},
-
-catTitle: {
-    fontSize: 22,
-    fontWeight: "bold",
-    marginLeft: 10,
-    color: "#1B263B",
-},
-
-catForm: {
-    backgroundColor: "#FFF",
-    padding: 20,
-    borderRadius: 16,
-    elevation: 3,
-    shadowColor: "#000",
-    shadowOpacity: 0.10,
-    shadowRadius: 4,
-    shadowOffset: { width: 0, height: 2 },
-},
-
-catLabel: {
-    fontSize: 16,
-    fontWeight: "600",
-    marginBottom: 8,
-    color: "#333",
-},
-
-catInput: {
-    backgroundColor: "#FFF",
-    padding: 14,
-    borderRadius: 10,
-    borderColor: "#DDD",
-    borderWidth: 1,
-    fontSize: 16,
-    marginBottom: 20,
-    elevation: 1,
-},
-
-catButton: {
-    backgroundColor: "#4C6EF5",
-    paddingVertical: 15,
-    borderRadius: 12,
-    alignItems: "center",
-},
-
-catButtonText: {
-    color: "#FFF",
-    fontSize: 18,
-    fontWeight: "bold",
-},
-// --- LIMITACIÓN DE GASTOS ---
-limitContainer: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: "#F5F5F5",
-},
-
-limitTitle: {
-    fontSize: 24,
-    fontWeight: "bold",
-    marginBottom: 20,
-    color: "#1B263B",
-},
-
-limitCard: {
-    backgroundColor: "#FFF",
-    padding: 20,
-    borderRadius: 16,
-    elevation: 4,
-    shadowColor: "#000",
-    shadowOpacity: 0.10,
-    shadowRadius: 5,
-    shadowOffset: { width: 0, height: 3 },
-},
-
-limitLabel: {
-    fontSize: 16,
-    fontWeight: "600",
-    marginTop: 15,
-    marginBottom: 6,
-    color: "#333",
-},
-
-limitInput: {
-    backgroundColor: "#FFF",
-    padding: 14,
-    borderRadius: 10,
-    borderWidth: 1,
-    borderColor: "#DDD",
-    fontSize: 16,
-    marginTop: 5,
-    marginBottom: 20,
-    elevation: 1,
-},
-
-limitOption: {
-    padding: 15,
-    borderWidth: 1,
-    borderColor: "#CCC",
-    borderRadius: 12,
-    marginTop: 10,
-    backgroundColor: "#FFF",
-    elevation: 1,
-},
-
-limitOptionText: {
-    fontSize: 16,
-    color: "#333",
-},
-
-limitBtnGuardar: {
-    backgroundColor: "#4C6EF5",
-    paddingVertical: 15,
-    borderRadius: 12,
-    alignItems: "center",
-    marginTop: 20,
-},
-
-limitTxtGuardar: {
-    color: "#FFF",
-    fontSize: 18,
-    fontWeight: "bold",
-},
-// --- DROPDOWN BONITO ---
-dropdownButton: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    backgroundColor: "#FFF",
-    borderWidth: 1,
-    borderColor: "#DDD",
-    padding: 14,
-    borderRadius: 10,
-    elevation: 2,
-},
-
-dropdownButtonText: {
-    fontSize: 16,
-    color: "#333",
-},
-
-dropdownContainer: {
-    marginTop: 6,
-    backgroundColor: "#FFF",
-    borderRadius: 10,
-    borderWidth: 1,
-    borderColor: "#DDD",
-    maxHeight: 240,
-    elevation: 4,
-    overflow: "hidden",
-},
-
-dropdownItem: {
-    padding: 14,
-    borderBottomWidth: 1,
-    borderBottomColor: "#EEE",
-},
-
-dropdownItemText: {
-    fontSize: 16,
-    color: "#333",
-},
-dropdownFlotante: {
-    position: "absolute",
-    top: 60,
-    left: 0,
-    right: 0,
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    paddingVertical: 6,
-    elevation: 10,
-    zIndex: 999,
-    maxHeight: 200, // Scroll interno
-    borderWidth: 1,
-    borderColor: "#ccc",
-},
-
-itemOpcion: {
-    paddingVertical: 12,
-    paddingHorizontal: 15,
-    borderBottomWidth: 1,
-    borderColor: "#eee",
-},
-
-itemOpcionTexto: {
-    fontSize: 16,
-},
-recientesWrapper: {
-    maxHeight: 750,        // Limita el alto
-    overflow: "hidden",    // Evita que el ScrollView se expanda
-    width: "100%",         // Ocupa todo el ancho disponible
-    marginBottom: 10,
-},
-
-
-
-
-
-
-
-});
