+++ styles/usuario.ts
@@ -1,0 +1,44 @@
+import { StyleSheet } from "react-native"
+
+export const usuario = StyleSheet.create({
+    loadingContainerusuario: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+    },
+
+    loadingTextusuario: {
+        fontSize: 18,
+    },
+    areausuario: {
+        padding: 20,
+        flex: 1,
+    },
+
+
+
+    titleusuario: {
+        fontSize: 20,
+        fontWeight: "bold",
+        marginBottom: 15,
+    },
+
+    itemusuario: {
+        marginBottom: 8,
+        fontSize: 16,
+    },
+
+    labelusuario: {
+        fontWeight: "bold",
+    },
+
+    backButton: {
+        width: 40,
+        height: 40,
+        borderRadius: 10,
+        backgroundColor: "#E9ECEF",
+        justifyContent: "center",
+        alignItems: "center",
+        marginBottom: 15,
+    },
+});
