+++ styles/vertodo.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native"
+
+export const vertodo = StyleSheet.create({
+backButton: {
+        width: 40,
+        height: 40,
+        borderRadius: 10,
+        backgroundColor: "#E9ECEF",
+        justifyContent: "center",
+        alignItems: "center",
+        marginBottom: 15,
+    },
+
+});
